requirements.txt
flake8==6.0.0
python-dotenv==1.0.0
requests==2.30.0
terminaltables==3.1.10
